requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
google-generativeai==0.8.3
groq==0.11.0
PyPDF2==3.0.1
pydantic==2.10.3